+++ include/LIEF/MachO/Binary.hpp
@@ -952,7 +952,7 @@ class LIEF_API Binary : public LIEF::Binary  {
   /// Check if the binary is supporting ARM64 pointer authentication (arm64e)
   bool support_arm64_ptr_auth() const {
     return header().cpu_type() == Header::CPU_TYPE::ARM64 &&
-           (header().cpu_subtype() & ~Header::CPU_SUBTYPE_MASK) == Header::CPU_SUBTYPE_ARM64_ARM64E;
+           (header().cpu_subtype() & ~Header::SUBTYPE_MASK) == Header::CPU_SUBTYPE_ARM64_ARM64E;
   }
 
   /// Return an iterator over the binding info which can come from either