+++ doc/sphinx/changelog.rst
@@ -162,7 +162,7 @@
 
   * Fix issue in the Python bindings while trying to access ``lief.__LIEF_MAIN_COMMIT__``
   * Fix CMake issue with ``find_package(lief-extended)``
-  * Use LLVM 20.1.3
+  * Use LLVM 20.1.6
 
 :Build System: